[feat] ModificationUI

+++ Assets/6.Debug/CustomDebugWindow.cs
@@ -1,4 +1,7 @@
-﻿namespace _6.Debug
+﻿using _1.Scripts.Manager.Core;
+using _1.Scripts.UI.InGame.Modification;
+
+namespace _6.Debug
 { 
 #if UNITY_EDITOR
     using System;
@@ -114,11 +117,39 @@
                 AssetDatabase.SaveAssets();
             }
 
+            if (GUILayout.Button("Get All Available Parts"))
+            {
+                playerObj = GameObject.FindWithTag("Player");
+                if (!playerObj.TryGetComponent(out Player player)) return;
+                
+                player.PlayerWeapon.GetAllAvailableParts();
+            }
+
+            if (GUILayout.Button("Forge Weapon"))
+            {
+                playerObj = GameObject.FindWithTag("Player");
+                if (!playerObj.TryGetComponent(out Player player)) return;
+
+                player.PlayerWeapon.ForgeWeapon();
+            }
+            
             if (GUILayout.Button("Clear PlayerPrefs"))
             {
                 PlayerPrefs.DeleteAll();
                 PlayerPrefs.Save();
                 Service.Log("PlayerPrefs 초기화 완료!");
+            }
+
+            if (GUILayout.Button("Show ModificationUI"))
+            {
+                CoreManager.Instance.uiManager.ShowUI<ModificationUI>();
+                Service.Log("ModificationUI Show");
+            }
+
+            if (GUILayout.Button("Hide ModificationUI"))
+            {
+                CoreManager.Instance.uiManager.HideUI<ModificationUI>();
+                Service.Log("ModificationUI Hide");
             }
         }
     }
